requirements.txt
Flask
PyYAML
python-dotenv
schedule
pytz
plexapi
dateutil